+++ core/src/main/kotlin/sift/core/element/AsmAliases.kt
@@ -1,0 +1,9 @@
+package sift.core.element
+
+typealias AsmType = org.objectweb.asm.Type
+typealias AsmAnnotationNode = org.objectweb.asm.tree.AnnotationNode
+typealias AsmClassNode = org.objectweb.asm.tree.ClassNode
+typealias AsmMethodNode = org.objectweb.asm.tree.MethodNode
+typealias AsmFieldNode = org.objectweb.asm.tree.FieldNode
+typealias AsmParameterNode = org.objectweb.asm.tree.ParameterNode
+